Streamlined files and patched jupyter vulnerability
