+++ app/lambda.py
@@ -1,9 +1,12 @@
 import logging
 import json
 import boto3
+import telebot
+
+import bot
 
 
-def handler(event: dict, context: dict) -> dict:
+def passthrough(event: dict, context: dict) -> dict:
     """
     This function replies to the gateway and passes on the event to the process_messages function.
 
@@ -21,3 +24,20 @@ def handler(event: dict, context: dict) -> dict:
         Payload=json.dumps(event),
     )
     return {"statusCode": 200}
+
+
+def converter(event: dict, context: dict) -> None:
+    """
+    This function handles messages from the user and directs them to the appropriate conversion function.
+
+    Args:
+        event: AWS Gateway event.
+        context: Lambda Context.
+
+    Returns: None.
+    """
+    logging.debug(f"Received event: {event}")
+    logging.debug(f"Received context: {context}")
+    update: telebot.types.Update = telebot.types.Update.de_json(json.loads(event['body']))
+    bot.tb.process_new_updates([update])
+    return
