--- package.txt
@@ -1,6 +1,0 @@
-python3
-python3-pip
-pandoc
-texlive-xetex
-texlive-fonts-recommended
-texlive-plain-generic
