--- scripts/set_webhook.sh
@@ -1,2 +1,0 @@
-curl https://api.telegram.org/bot"${TELEGRAM_TOKEN}"/setWebhook?url="$1"
-echo "/n"
